feat(web): make Mira the workspace dashboard
