+++ apps/docs/platform/architecture/tanstack-rust-migration-plan.mdx
@@ -119,7 +119,7 @@ handoffs visible so parallel workers do not repeat the same audit:
 | `/:locale/:wsId/users/groups/:groupId/attendance` | Migrated in the TanStack frontend as a dynamic no-store page with fail-closed auth, workspace and group checks, `check_user_attendance`, `update_user_attendance` save gating, session/date URL state, and typed internal-api attendance/member/config/session facades. | Backend agents still need Rust ownership for the user-group attendance, members, workspace config, and session APIs before full backend cutover. |
 | `/:locale/:wsId/billing` | `apps/pay` owns detailed billing, checkout, invoices, credits, subscriptions, webhooks, and payment mutations. `apps/web` retains only a read-only summary loaded from Pay through `packages/internal-api`; all actions link to Pay. | Treat the remaining web summary as an intentional satellite-app handoff. Do not restore payment mutations or detailed billing UI to either platform frontend. |
 | `/:locale/:wsId/usage` | The page wrapper is small, but the content is a broad Supabase/RPC workspace usage dashboard. | Create a usage-summary backend contract or server-function bundle that returns permission-masked stats without direct Supabase reads in TanStack. |
-| `/:locale/:wsId/ai/spark`, `/:locale/:wsId/mira`, `/:locale/:wsId/assistant` | Page files are compact, but the behavior is coupled to AI object APIs, Mira API routes, voice/live API helpers, or assistant providers. | Treat these as dedicated product lanes with typed AI/Mira/assistant facades and focused browser evidence. |
+| `/:locale/:wsId/ai/spark`, `/:locale/:wsId/mira`, and the canonical `/:locale/:wsId` Mira dashboard | Page files are compact, but the behavior is coupled to AI object APIs, Mira API routes, voice/live API helpers, or assistant providers. The standalone `/:locale/:wsId/assistant` route has been removed in favor of the unified workspace-root chat and rich voice experience. | Treat these as dedicated product lanes with typed AI/Mira/assistant facades and focused browser evidence. |
 
 ## Coordinator-Owned Artifacts
